agents/calendar_sync/requirements.txt
msal==1.31.0
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
pydantic==2.10.3
python-dotenv==1.0.1
pyyaml==6.0.2
